fallout: fixer for ncr cop aggro
Adds a CLI tool to fix aggro for NCR cops. You'll need to fix the saves
one by one, but it should still work.

The code behind it is probably the worst I've ever written. I just
wanted to get this part done quickly, as it was my first goal.