rojekti: parse window config
Handles single command configuration as with serde-yaml. Fixes incorrect
handling of the 'null' value. It is a scalar value in yaml and signifies
the absence of value. You live and learn.

Only the 'nil' value is not parsed as a program. If I remember correctly
the 'nil' was supported in tmuxinator because of the dynamism in Ruby.
They parsed values would translate to ruby objects and nil would become
null as well.