swkotor: basic dll build support
Using crane and fenix we can easily cross compile a dll. I've only
tested that the project does build a dll but not that if the dll is in
any shape or form usable. Main goal was to get nix setup for cross
compilation.

Next up is testing if we can load this dll under wine for kotor.