util: add simple poc tools to handle memory
Tools meant as PoC so far to handle memory in a process. The tools are
crude and presumably should deprecate before any serious use.

In order to be able to edit memory, add memory_patcher module to replace
memory content with given data.

To have something concrete to edit, add needle_finder module to search
for a specific byte sequence in the process memory. To be able to test
it out, add a few simple poc functions for the sake of dev-verbosity.