swkotor: add launcher script
Nixified launcher script for starting the game through terminal.

The steam web protocol arguments don't seem to be the same as launch
options. Launch options can affect how the whole executable is started,
through use of %command%. Where the launch options just seem to be
arguments passed down to the executable.

I'll have to ask Valve if this is even supported or if there are other
options to do this.