qa: add live quality assurance test framework
In order to be able to confirm that whatever is made on one platform
works on the other, add a simple framework to run tests live when the
dll is loaded.

Normal cargo tests cannot be performed as many of the ones that will be
tested by this framework will not be usable in a normal test
environment, as they will need to be run in a live environment.

See the examples/rustdoc in liveqa/runner.rs and liveassert.rs.