tests: add tests for iat replacing
In order to be sure that IAT replacments work, add tests for it. The
tests will try to replace a CreateFileA hook and the process what was
done.

Unfortunately, rust will not allow using closures with `extern "system"`
function, but we'll make do with static ones defined in the functions.