crago: add plthook dependency
`plthook` will be used in a future commit to get address of functions in
different windows modules (DLL or EXE memoryspaces). Add it beforehand
to be clear about the edit to dependencies.

For the record, normally this might not be warranted. But in this case,
the dependencies added to Cargo.lock are too much with the evenual
change that uses the `plthook`.