iat: add possibilty to do iat hooks
IAT (Import Access Table) hooks are a way to hook into the import table
of an EXE or DLL memory space and replace the function calls with your
own. These are similar to trampoline hooks, but are more powerful and
easier to use as the information is already available in the PE/DLL
import table.

The module is tried to be made as generic as possible, but the only
when another function is used, it will be seen if it is usable enough.
On top of that, the current definition in createfile.rs is hard to
grasp, and could use some helper functions.

It is not guaranteed that IAT will be needed, but it is a useful utility
to have.

As a proof-of-concept, add a createfile hook. It will log if
"dialog.tlk" file is accessed. It might be needed in the future to hook
into if the tlk will be patched without edits to the file.