iat: add hook to swapbuffers
`SwapBuffers` is a function that windows calls after each frame. Note
that sometimes it's the same as wglSwapBuffers. This function is usable
to handle custom rendering after the game has done its work. When
called, any mod-rendered item will/can be shown on top.

To be able to use it, add an open callback function that can be filled
out. This way, the heavy liftin will be up to the one that sets up the
function call.