graphics: add routine to build opengl binding
SWKotor uses OpenGL version 2.1 (As best interpreted). To be able to
call functions from that era, we need to be able to generate the
bindings for that specific version. See [1] for more information.

Note that this commit alone does not enable the functions to be called.
They still need to be setup with `gl::load_with`. This will be done in
a future commit.

[1]: https://docs.rs/gl_generator/latest/gl_generator/