graphics: draw some boxes to top left
Mainly as a PoC (while not in poc directory), draw some boxes in the
top left corner. No idea if this will work in all situations, but at
least now it's stored properly.

Note that the location for the draw function is poor. It is among
setup code. It should be edited so that it will be called on demand,
preferrably by another callback.