flake: find nix projects automatically
Detect presence of `package.nix` in a subdirectory and load it. Removing
the need to define each project by hand.

Use `callPackageWith` to create a custom `callPackage` with `inputs` in
scope to remove need for manually passing inputs to projects.

I'm still on the fence if this is just utter wankery and will this make
it harder to reason about projects. The intend is to merge `package.nix`
and `shell.nix` within a project to one file and pass these to the root
flake. Having this automatic wankery should make that setup simpler,
with the cost of making code harder to read: Nix REPL is probably the
only sane way to reason at that point.